pyproject.toml
[tool.poetry]
name = "blog-drf"
version = "0.1.0"
description = ""
authors = ["Sviatoslav Bobryshev <s.bobryshev@tg.dunice.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
psycopg = {extras = ["binary"], version = "^3.1.11"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.291"
black = "^23.9.1"


[tool.poetry.group.test.dependencies]
pytest = {extras = ["cov"], version = "^7.4.2"}
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"
ipython = "^8.15.0"
pysnooper = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "blog_drf.settings"
